// src/shared/source-wikimapia/types.ts
import * as turf from "@turf/turf";

import { Tile } from "../tiles";

export type ProcessedWikimapiaTileResponse = Array<
  turf.Feature<turf.GeometryCollection>
>;

export interface WikimapiaTileData {
  tile: Tile;
  fetchedAt: string;
  response: ProcessedWikimapiaTileResponse;
}

export type WikimapiaObjectPointFeature = turf.Feature<
  turf.Point,
  {
    description?: string;
    wikimapiaId: number;
  }
>;

export type WikimapiaObjectExtentFeature = turf.Feature<
  turf.Polygon,
  {
    description?: string;
    wikimapiaId: number;
  }
>;

export type WikimaiaTileExtentFeature = turf.Feature<
  turf.Polygon,
  {
    tileId: string;
    fetchedAt: string;
    fetchedFeatureCount: number;
  }
>;

export interface WikimapiaObjectPhotoInfo {
  url: string;
  userId: number;
  userName: string;
}

export interface WikimapiaObjectInfo {
  wikimapiaId: number;

  /** @deprecated Used before 2021-04-05 */
  completionDates?: string;
  completionTime?: string;
  name?: string;
  photos?: WikimapiaObjectPhotoInfo[];
  demolished?: boolean;
}

export interface WikimapiaObjectInfoFile {
  fetchedAt: string;
  parsedAt: string;
  data: WikimapiaObjectInfo;
}